permuter sweep 95-99% band: 18 wins, +30.8% total match
batch_auto resumable sweep over 487 AT_LIMIT functions in the 95-99%
band. 24 functions improved, 18 retained here (6 zero-delta entries
reverted as churn without measurable benefit, per the FftIpp lesson).

Two perfects (100%) — verification pending (build manifest broke
mid-session from a date jump; will re-confirm next session):
  Memory_Xbox PhysicalAlloc        98.7 -> 100.0   (+1.3)
  rndobj/Graph RndGraph::DrawAll   96.1 -> 100.0   (+3.9)

Top remaining wins by delta:
  hamobj/HamWardrobe OnAddCrowd            78.3 -> 91.6   (+13.3)
  net/DingoSvr DingoServer::AddDelayedCalls 96.7 -> 99.9  (+3.2)
  utl/Cache_Xbox CacheXbox::DeleteParentDirs 98.3 -> 99.9 (+1.6)
  hamobj/MoveAsyncDetector EnableDetector  92.4 -> 93.5   (+1.1)
  lazer/game/PartyModeMgr ReadPartySongQueue 98.3 -> 99.3 (+1.0)
  rndobj/Text UpdateScrollOffsets           97.4 -> 98.4  (+1.0)
  rndobj/PropKeys Vector3Keys::SetFrame     98.4 -> 99.3  (+0.9)
  hamobj/HamDirector DrawIconMan            97.8 -> 98.7  (+0.9)
  flow/FlowTrigger FlowTrigger::Load        98.1 -> 98.8  (+0.7)
  rndobj/Shader CalcShaderOpts              97.6 -> 98.1  (+0.5)
  char/CharIKHand Highlight                 96.9 -> 97.4  (+0.5)
  world/CameraShot BuildTransform           97.2 -> 97.5  (+0.3)
  meta_ham/KinectSharePanel ConvertImages   97.3 -> 97.5  (+0.2)
  keygen_xbox parseHex16                    96.4 -> 96.6  (+0.2)
  rndobj/HiResScreen Merge                  97.5 -> 97.6  (+0.1)
  char/ClipDistMap CalcWidth                96.5 -> 96.6  (+0.1)

Auto-generated _tmp0/_tmp1 names renamed to meaningful identifiers
(physSize, nextIdx, end, moveName, msg) — names don't affect codegen.
FlowTrigger's auto-expanded foreach->do-while indentation cleaned up
and a comment added explaining why the manual loop matches the target.

+++ src/system/char/ClipDistMap.cpp
@@ -321,7 +321,7 @@ int ClipDistMap::CalcWidth() {
 
     float aStart = mAStart;
     int spb = mSamplesPerBeat;
-    int width = Max(0, (int)(float)floor(((mAEnd - aStart) * (float)spb) + 0.5f)) + 1;
+    int width = Max(0, (int)(float)floorf(((mAEnd - aStart) * (float)spb) + 0.5f)) + 1;
     mAEnd = mAStart + (float)(width - 1) / (float)mSamplesPerBeat;
     return width;
 }
